+++ R/ct_fit_activity.R
@@ -4,7 +4,7 @@
 #'
 #' @param time_of_day A numeric vector of radian time-of-day data
 #' @param weights A numeric vector of weights for each dat value.
-#' @param n_boostrap Number of bootstrap iterations to perform. Ignored if sample=="none"
+#' @param n_bootstrap Number of bootstrap iterations to perform. Ignored if sample=="none"
 #' @param bandwidth Numeric value for kernel bandwidth. If NULL, calculated internally.
 #' @param adjustment Numeric bandwidth adjustment multiplier.
 #'
@@ -23,7 +23,7 @@
 #'
 #'
 #' fit_act <- ct_fit_activity(time_of_day = observations$time_radian,
-#'                            sample = "model", n_boostrap = 100)
+#'                            sample = "model", n_bootstrap = 100)
 #'
 #' # Access activity level estimation
 #' fit_act$activity
@@ -31,16 +31,18 @@
 #' @export
 ct_fit_activity <- function(time_of_day,
                             weights = NULL,
-                            n_boostrap = 1000,
+                            n_bootstrap = 1000,
                             bandwidth  = NULL,
                             adjustment = 1,
                             sample = c("none", "data", "model"),
                             bounds = NULL,
                             show = TRUE
                             ) {
+  sample <- match_arg(sample, choices = c("none", "data", "model"))
+
   fit_act <- activity::fitact(dat = time_of_day,
                               wt = weights,
-                              reps = n_boostrap,
+                              reps = n_bootstrap,
                               bw = bandwidth,
                               adj = adjustment,
                               sample = sample,
@@ -53,9 +55,13 @@ ct_fit_activity <- function(time_of_day,
     bandwidth = fit_act@bw,
     adjustement = fit_act@adj,
     pdf = dplyr::as_tibble(fit_act@pdf),
-    activity = t(fit_act@act) %>%
-      as.data.frame() %>% dplyr::as_tibble() %>%
-      dplyr::rename(lower_ci = 3, upper_ci = 4)
+    activity = {
+      act <- t(fit_act@act) %>%
+        as.data.frame() %>% dplyr::as_tibble()
+      if (any(grepl('lcl', names(act)))) {
+        act <- act %>% dplyr::rename(lower_ci = 3, upper_ci = 4)
+      }
+    }
   ))
 }